refactoring union decorator structure
