+++ src/index.ts
@@ -1,4 +1,6 @@
 export * from './decorator';
+export * from './decorator/';
+export * from './metadata/options';
 export { schemaFactory } from './schema_factory';
 export * from './order-by-item';
 export * from './pagination.type';
