+++ src/metadata/types/index.ts
@@ -1,0 +1,1 @@
+export * from './union.metadata';
