+++ src/metadata/types/union.metadata.ts
@@ -1,5 +1,5 @@
 export interface UnionTypeMetadata {
-    name?: string;
+    name: string;
     description?: string;
     types: any[];
     resolver: (obj: any, context: any, info: any) => Promise<string> | string | null;
